Move maskBracketForDate to bracket-data module
Extract from tournament-app.ts to bracket/bracket-data.ts for testability.
Also add score annotation policy to CLAUDE.md (from #157).

Refs #153
